Fixed a bug when "UseSquarePixels=false"
When the "UseSquarePixels" setting is set to false, then the program does not attempt to maintain the proper aspect ratio. Windowed width/height is used as-is, with no correction. However, there was a bug when going into fullscreen mode where the "UseSquarePixels" property was not being obeyed. This has now been fixed.